requirements.txt
torch==1.10.2
torchvision==0.11.3
timm==0.3.2
transformers==2.0.1
numpy==1.18.5
pillow==9.10
google-cloud-storage